ATE_LOOP_COUNT_2 should be i8's not u64's

+++ ec/src/models/mnt4/mod.rs
@@ -25,7 +25,7 @@ pub type GT<P> = Fp4<P>;
 pub trait MNT4Parameters: 'static {
     const TWIST: Fp2<Self::Fp2Config>;
     const TWIST_COEFF_A: Fp2<Self::Fp2Config>;
-    const ATE_LOOP_COUNT_2: &'static [u64];
+    const ATE_LOOP_COUNT_2: &'static [i8];
     const ATE_IS_LOOP_COUNT_NEG: bool;
     const FINAL_EXPONENT_LAST_CHUNK_1: <Self::Fp as PrimeField>::BigInt;
     const FINAL_EXPONENT_LAST_CHUNK_W0_IS_NEG: bool;
